+++ packages/components/organisms/f-checkout/test/component/f-checkout.component.shared.spec.js
@@ -1,6 +1,5 @@
 const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions.js');
 const Checkout = require('../../test-utils/component-objects/f-checkout.component');
-
 let checkout;
 
 describe('f-checkout component tests - @browserstack', () => {
@@ -14,11 +13,6 @@ describe('f-checkout component tests - @browserstack', () => {
 
         checkout.open(pageUrl);
         checkout.waitForComponent();
-    });
-
-    it('should display the f-checkout component', () => {
-        // Assert
-        expect(checkout.isComponentDisplayed()).toBe(true);
     });
 
     it.skip('should submit the checkout form', () => {
@@ -39,50 +33,6 @@ describe('f-checkout component tests - @browserstack', () => {
 
         // Assert
         // Waiting for route here, so we can grab redirect url and show form submits.
-    });
-
-    it('should display the mandatory "mobileNumber" field', () => {
-        // Assert
-        expect(checkout.isFieldDisplayed('mobileNumber')).toBe(true);
-    });
-
-    it('should display a "mobileNumber" error message when an unsupported country code is used in the mobile number field', () => {
-        // Arrange
-        const addressDetails = {
-            mobileNumber: '+8112345678911'
-        };
-
-        // Act
-        checkout.populateCheckoutForm(addressDetails);
-        checkout.goToPayment();
-
-        // Assert
-        expect(checkout.isFieldErrorDisplayed('mobileNumber')).toBe(true);
-    });
-
-    it('should not display a "mobileNumber" error message when a number is formatted with a supported country code', () => {
-        // Arrange
-        const addressDetails = {
-            mobileNumber: '+4412345678911'
-        };
-
-        // Act
-        checkout.populateCheckoutForm(addressDetails);
-        checkout.goToPayment();
-
-        // Assert
-        expect(checkout.isFieldErrorDisplayed('mobileNumber')).toBe(false);
-    });
-
-    it('should display times in ascending order, with default text "As soon as possible" showing first', () => {
-        // Act
-        checkout.selectOrderTime('As soon as possible');
-
-        // Assert
-        expect(checkout.isOrderTimeDropdownDisplayed()).toBe(true);
-        expect(checkout.getOrderTimeOptionText(0)).toBe('As soon as possible');
-        expect(checkout.getOrderTimeOptionText(1)).toBe('Wednesday 01:45');
-        expect(checkout.getOrderTimeOptionText(2)).toBe('Wednesday 02:00');
     });
 
     it('should prevent a user from writing a note of over 200 characters', () => {
@@ -119,51 +69,6 @@ describe('f-checkout component tests - @browserstack', () => {
         // Waiting for route here, so we can grab redirect url and show form submits.
     });
 
-    it('should display the switch user link', () => {
-        expect(checkout.switchUserLinkIsDisplayed()).toBe(true);
-    });
-
-    it('should display the preorder warning message when ASAP is not avalible', () => {
-        // Arrange
-        checkout.withQuery('&knob-Is ASAP available', false);
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
-        // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
-
-        // Assert
-        expect(checkout.isPreOrderWarningDisplayed()).toBe(true);
-    });
-
-
-    it('should display the checkout error component when "Has Checkout Errors" is true', () => {
-        // Arrange
-        checkout.withQuery('&knob-Checkout Errors', 'ISSUES');
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
-        // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
-        checkout.goToPayment();
-
-        // Assert
-        expect(checkout.isCheckoutErrorMessageDisplayed()).toBe(true);
-    });
-
-    it('should display the checkout error component when "Has Place Order Errors" is true', () => {
-        // Arrange
-        checkout.withQuery('&knob-Place Order Errors', 'SERVER');
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
-        // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
-        checkout.goToPayment();
-
-        // Assert
-        expect(checkout.isCheckoutErrorMessageDisplayed()).toBe(true);
-    });
 
     it('should close the checkout error when "Retry" is clicked', () => {
         // Arrange