+++ packages/components/organisms/f-cookie-banner/test/component/f-cookieConsentBanner.component.desktop.spec.js
@@ -28,7 +28,6 @@ describe('New - f-cookieBanner component tests - @browserstack', () => {
         // Assert
         expect(bannerCookie.value).toBe('130315');
         expect(bannerConsent.value).toBe(expectedCookieValue);
-        expect(cookieBanner.isCookieBannerComponentDisplayed()).toBe(false);
     });
 });