+++ packages/components/organisms/f-footer/test/component/f-footer.component.desktop.spec.js
@@ -77,13 +77,14 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(browser.getUrl()).toContain(expectedUrl);
     });
 
-    it('should display the footer', () => {
+    it('should display the footer - @percy', () => {
         // Assert
         expect(footer.isComponentDisplayed()).toBe(true);
     });
 
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['gb', 'au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it('should not show courier links and country selector for country code "%s" when options are unselected', expectedLocale => {
+    .it('should not show courier links and country selector for country code "%s" when options are unselected - @percy', expectedLocale => {
         // Arrange
         footer.withQuery('&knob-Locale', formatLocale(expectedLocale));
         footer.withQuery('&knob-Show country selector', 'false');
@@ -99,8 +100,9 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(footer.isCountrySelectorDisplayed()).toBe(false);
     });
 
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['au', 'ie', 'nz'])
-    .it('should show courier links for country code "%s" when option is selected', expectedLocale => {
+    .it('should show courier links for country code "%s" when option is selected - @percy', expectedLocale => {
         // Arrange
         footer = new Footer('organism', 'footer-component');
         footer.withQuery('&knob-Locale', formatLocale(expectedLocale));
@@ -116,8 +118,9 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(footer.areCourierLinksDisplayed()).toBe(true);
     });
 
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['gb', 'es', 'it', 'no'])
-    .it('should never show courier links for country code "%s", even when option is selected', expectedLocale => {
+    .it('should never show courier links for country code "%s", even when option is selected - @percy', expectedLocale => {
         // Arrange
         footer = new Footer('organism', 'footer-component');
         footer.withQuery('&knob-Locale', formatLocale(expectedLocale));
@@ -133,8 +136,9 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(footer.areCourierLinksDisplayed()).toBe(false);
     });
 
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it.only('should always show country selector for country code "%s" when selected', expectedLocale => {
+    .it('should always show country selector for country code "%s" when selected - @percy', expectedLocale => {
         // Arrange
         footer = new Footer('organism', 'footer-component');
         footer.withQuery('&knob-Locale', formatLocale(expectedLocale));
@@ -150,8 +154,9 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(footer.isCountrySelectorDisplayed()).toBe(true);
     });
 
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['au', 'gb', 'nz', 'ie', 'dk', 'es', 'it'])
-    .it('should display the corresponding icon for the "%s" country code', expectedLocale => {
+    .it('should display the corresponding icon for the "%s" country code - @percy', expectedLocale => {
         // Arrange
         footer = new Footer('organism', 'footer-component');
         footer.withQuery('&knob-Locale', formatLocale(expectedLocale));
@@ -221,9 +226,9 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(browser.getUrl()).toContain(expectedUrl);
     });
 
-
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['gb', 'au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it('should display social icons block for country code "%s"', expectedLocale => {
+    .it('should display social icons block for country code "%s" - @percy', expectedLocale => {
         // Arrange
         footer = new Footer('organism', 'footer-component');
         footer.withQuery('&knob-Locale', formatLocale(expectedLocale));
@@ -239,8 +244,9 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(footer.isSocialIconBlockDisplayed()).toBe(true);
     });
 
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['gb', 'au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it('should display app downloads block for country code "%s"', expectedLocale => {
+    .it('should display app downloads block for country code "%s" - @percy', expectedLocale => {
         // Arrange
         footer = new Footer('organism', 'footer-component');
         footer.withQuery('&knob-Locale', formatLocale(expectedLocale));
@@ -256,8 +262,9 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(footer.isDownloadIconBlockDisplayed()).toBe(true);
     });
 
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['gb', 'au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it('should display payment options block for country code "%s"', expectedLocale => {
+    .it('should display payment options block for country code "%s" - @percy', expectedLocale => {
         // Arrange
         footer = new Footer('organism', 'footer-component');
         footer.withQuery('&knob-Locale', formatLocale(expectedLocale));
@@ -273,8 +280,9 @@ describe('Desktop - f-footer component tests - @browserstack', () => {
         expect(footer.isPaymentIconsBlockDisplayed()).toBe(true);
     });
 
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['gb', 'au', 'ie', 'nz', 'dk', 'es', 'it', 'no'])
-    .it('should display the feedback block for country code "%s"', expectedLocale => {
+    .it('should display the feedback block for country code "%s" - @percy', expectedLocale => {
         // Arrange
         footer = new Footer('organism', 'footer-component');
         footer.withQuery('&knob-Locale', formatLocale(expectedLocale));