+++ packages/components/organisms/f-header/test/component/f-header.component.mobile.spec.js
@@ -51,14 +51,15 @@ describe('Mobile - f-header component tests - @browserstack', () => {
         }
     });
 
+    // Refactor for Percy visual regression
     forEach(['offersLink', 'delivery', 'help', 'countrySelector', 'userAccount'])
-    .it('should open navigation and display link "%s"', link => {
+    .it('should open navigation and display link "%s" - @percy', link => {
         // Assert
         expect(header.isNavigationLinkDisplayed(link)).toBe(true);
     });
 
     forEach(['au', 'gb', 'nz', 'ie', 'dk', 'es', 'it'])
-    .it('should open country selector and display country code "%s"', country => {
+    .it('should open country selector and display country code "%s" - @percy', country => {
         // Act
         header.openCountrySelector();
         header.expectedCountry = country;
@@ -67,8 +68,9 @@ describe('Mobile - f-header component tests - @browserstack', () => {
         expect(header.isCountryLinkDisplayed()).toBe(true);
     });
 
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['au', 'ie', 'nz'])
-    .describe('closed navigation for country code "%s"', expectedLocale => {
+    .describe('closed navigation for country code "%s" - @percy', expectedLocale => {
 
         beforeEach(() => {
             // Act
@@ -81,19 +83,19 @@ describe('Mobile - f-header component tests - @browserstack', () => {
         });
 
         forEach(['offersLink', 'delivery', 'userAccount', 'help', 'countrySelector'])
-        .it('should not display "%s" navigation link', link => {
+        .it('should not display "%s" navigation link - @percy', link => {
             // Assert
             expect(header.isNavigationLinkDisplayed(link)).toBe(false);
         });
 
-        it('should display offersIcon link', () => {
+        it('should display offersIcon link - @percy', () => {
             // Assert
             expect(header.isNavigationLinkDisplayed('offersIcon')).toBe(true);
         });
     });
 
     forEach(['au', 'ie', 'nz'])
-    .describe('open navigation for country code "%s"', expectedLocale => {
+    .describe('open navigation for country code "%s" - @percy', expectedLocale => {
         beforeEach(() => {
             // Act
             header.withQuery('&knob-Locale', formatLocale(expectedLocale));
@@ -106,19 +108,19 @@ describe('Mobile - f-header component tests - @browserstack', () => {
         });
 
         forEach(['offersLink', 'userAccount', 'help', 'countrySelector'])
-        .it('should display navigation link "%s"', link => {
+        .it('should display navigation link "%s" - @percy', link => {
             // Assert
             expect(header.isNavigationLinkDisplayed(link)).toBe(true);
         });
 
-        it('should not display offersIcon link', () => {
+        it('should not display offersIcon link - @percy', () => {
             // Assert
             expect(header.isNavigationLinkDisplayed('offersIcon')).toBe(false);
         });
     });
 
     forEach(['it', 'es', 'dk', 'no'])
-    .describe('closed navigation for country code "%s"', expectedLocale => {
+    .describe('closed navigation for country code "%s" - @percy', expectedLocale => {
         beforeEach(() => {
             // Act
             header.withQuery('&knob-Locale', formatLocale(expectedLocale));
@@ -130,18 +132,19 @@ describe('Mobile - f-header component tests - @browserstack', () => {
         });
 
         forEach(['offersLink', 'delivery', 'userAccount', 'help', 'countrySelector'])
-        .it('should not display navigation link "%s",', link => {
+        .it('should not display navigation link "%s" - @percy', link => {
             // Assert
             expect(header.isNavigationLinkDisplayed(link)).toBe(false);
         });
 
-        it('should not display the offersIcon', () => {
+        it('should not display the offersIcon - @percy', () => {
             expect(header.isNavigationLinkDisplayed('offersIcon')).toBe(false);
         });
     });
 
+    // Make sure tenant is appended to screenshot for Percy tests
     forEach(['it', 'es', 'dk', 'no'])
-    .describe('open navigation for country code "%s"', expectedLocale => {
+    .describe('open navigation for country code "%s" - @percy', expectedLocale => {
         beforeEach(() => {
             // Act
             header.withQuery('&knob-Locale', formatLocale(expectedLocale));
@@ -155,22 +158,22 @@ describe('Mobile - f-header component tests - @browserstack', () => {
 
 
         forEach(['userAccount', 'help', 'countrySelector'])
-        .it('should display navigation link "%s"', link => {
+        .it('should display navigation link "%s" - @percy', link => {
             // Assert
             expect(header.isNavigationLinkDisplayed(link)).toBe(true);
         });
 
-        it('should not display the offers link', () => {
+        it('should not display the offers link - @percy', () => {
             // Assert
             expect(header.isNavigationLinkDisplayed('offersLink')).toBe(false);
         });
 
-        it('should not display the delivery link', () => {
+        it('should not display the delivery link - @percy', () => {
             // Assert
             expect(header.isNavigationLinkDisplayed('delivery')).toBe(false);
         });
 
-        it('should not display the offersIcon link', () => {
+        it('should not display the offersIcon link - @percy', () => {
             // Assert
             expect(header.isNavigationLinkDisplayed('offersIcon')).toBe(false);
         });