+++ test/configuration/chrome/chrome.settings.js
@@ -39,6 +39,18 @@ const settings = () => ({
             }] : [])
         ]
     },
+    visual: {
+        capabilities: [
+            {
+                browserName: 'chrome',
+                acceptInsecureCerts: true,
+                specs: [
+                    'test/visual/*.visual.desktop.spec.js',
+                    'test/visual/*.visual.mobile.spec.js',
+                    'test/visual/*.visual.shared.spec.js'
+                ]
+            }
+        ]},
     reporters: ALLURE_REPORTER === 'true' ? [
         [video, {
             saveAllVideos: false, // If true, also saves videos for successful test cases