+++ test/utils/configuration-helper.js
@@ -1,9 +1,22 @@
-const { CIRCLECI } = process.env;
+const { CIRCLECI, JE_ENV, TEST_TYPE } = process.env;
+const chromeSettings = require('../configuration/chrome/chrome.settings').default();
+const browserstackSettings = require('../configuration/browserstack/browserstack.settings').default();
+const environmentSettings = CIRCLECI ? require('../configuration/ci.test-settings').default()
+: require('../configuration/local.test-settings').default()
+const isLocal = JE_ENV === 'local';
 
-exports.setTestSettings = () => {
+exports.getConfigurationSettings = () => {
 
-    const testSettings = CIRCLECI ? require('../configuration/ci.test-settings').default()
-                                    : require('../configuration/local.test-settings').default()
+    let runnerSettings = isLocal ? chromeSettings : browserstackSettings;
+    let testCapabilities = runnerSettings[TEST_TYPE.toLowerCase()].capabilities;
 
-    return testSettings;
+    return {
+        capabilities: testCapabilities,
+        bail: environmentSettings.bail,
+        logLevel: environmentSettings.logLevel,
+        maxInstances: environmentSettings.maxInstances, 
+        baseUrl: runnerSettings.baseUrl,
+        reporters: runnerSettings.reporters,
+        services: runnerSettings.services,
+    }
 };