+++ packages/components/organisms/f-cookie-banner/CHANGELOG.md
@@ -4,11 +4,26 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-Latest to be added to the next release
+v0.9.0
 ------------------------------
+*May 10, 2021*
+
+### Removed
+- `is-visuallyHidden` css style definition as should come from `fozzie`
+
+
+v0.8.0
+------------------------------
+*May 6, 2021*
+
+### Added
+- `f-mega-modal` component to cookie banner.
+
+### Removed
+- Styles related to old modal & modal container.
 
 ### Fixed
-- Added documentation to indicate cookie-banner will delete cookies conditionally that are not in the exclusion list
+- Added documentation to indicate cookie-banner will delete cookies conditionally that are not in the exclusion list.
 
 
 v0.7.1