+++ packages/components/organisms/f-footer/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v4.20.0
+------------------------------
+*May 10, 2021*
+
+### Added
+- Tagging for future Percy tests
+- euro2020 related links to be shown if `euro2020Enabled` is true
+
+
 v4.19.0
 ------------------------------
 *April 15, 2021*