+++ packages/components/organisms/f-header/CHANGELOG.md
@@ -3,6 +3,19 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v4.15.0
+------------------------------
+*May 6, 2021*
+
+### Removed
+- `is-visuallyHidden`, `is-hidden`, `is-hidden--noJS`, `is-shown`, `is-shown--noJS` css style definition as should come from `fozzie`
+
+### Added
+- Focus styles for navigation links for transparent and orange header versions
+- Tagging for future Percy tests
+
+
 v4.14.0
 ------------------------------
 *May 4, 2021*