+++ packages/components/organisms/f-registration/CHANGELOG.md
@@ -3,12 +3,25 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v0.52.0
+------------------------------
+*May 6, 2021*
+
+### Removed
+- `is-visuallyHidden` css style definition as should come from `fozzie`
+
+### Added
+- Tagging for future Percy tests
+
+
 v0.51.0
 ------------------------------
 *April 15, 2021*
 
 ### Changes
 - Error messages retrieved from i18n files.
+
 
 v0.50.0
 ------------------------------